backend/requirements.txt
alembic==1.14.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
Mako==1.3.8
MarkupSafe==3.0.2
pycparser==2.22
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
